+++ localvoice/lang/__init__.py
@@ -18,7 +18,7 @@ import importlib
 import pkgutil
 
 REQUIRED = ("CODE", "PATTERNS", "NUM_WORDS", "ORDINAL_WORDS",
-            "MINUTE_WORDS", "DURATIONS", "MOOD_WORDS")
+            "MINUTE_WORDS", "DURATIONS", "MOOD_WORDS", "PLAY_VERBS")
 
 #: ``{"it": <module lang.it>, "en": <module lang.en>, ...}``
 PACKS = {}